Moved `deviceName` from nonTandemData to general data util
